requirements.txt
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.2
openai>=1.0.0
pinecone>=7.3.0
python-dotenv>=1.0.0
markdown2>=2.4.10
selenium>=4.21.0
webdriver-manager>=4.0.1
